+++ test/fixtures/cookbooks/summary_handlers_test/metadata.rb
@@ -1,0 +1,9 @@
+name 'summary_handlers_test'
+maintainer 'ChrisGit'
+maintainer_email ''
+license 'all_rights'
+description 'Show Typical Output For Summary Handlers'
+long_description 'Summary Handlers Test Cookbook'
+version '1.0.0'
+
+supports 'windows'