+++ test/fixtures/cookbooks/summary_handlers_test/recipes/default.rb
@@ -1,0 +1,8 @@
+#
+# Cookbook Name:: summary_handlers_test
+# Recipe:: default
+#
+# Copyright (c) 2017 ChrisGit, All Rights Reserved.
+
+include_recipe 'summary_handlers_test::simple_config'
+include_recipe 'summary_handlers_test::web_site'